[Bridge 18/n] add bridge move package as an example (#15490)
## Description 

This is basically a copy of https://github.com/MystenLabs/sui/pull/15411
which is very hard to integrate with until it's merged into main. At the
moment we want more eyes to review it. This PR enables programatic e2e
testing.

Code added in this PR will be cleaned up eventually once the
aforementioned PR is merged.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes