[move] Add freezing for tail expressions (#15588)
## Description 

There was a small oversight in freezing bound expressions in tail
positions introduced during the HLIR rewrite. The VM validator caught
the issue, but it would be nice to allow this code since it's valid Move.

## Test Plan 

Added a new test file to cover the behavior.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes