move-package: use original package names in resolution error messages (#15564)
## Description 

When using pkg name resolution, dependency error messages during graph
resolution will print resolved names instead of the original ones as
defined in manifests which will be confusing to the user. This commit
makes it so that original package names are used. This is achieved by
storing the original name in the `Dependency` struct which is the edge
between two packages in the graph. This way the original name can always
be recovered. This commit should only affect the printed error messages
(and not functionality of the toolchain).

This is part of the work to enable compiling against on-chain
dependencies.

## Test Plan 

Unit tests reflect the changes

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes