[GraphQL/Cursor] Generalise cursor type used by Page
`Page<C>` was originally designed to work with `Cursor<C>`. This
change allows it to be used with any type that implements
`CursorType`.

The motivation for this is the upcoming port of `Object` to the cursor
and pagination framework: We need to be able to use `Page` to paginate
Objects, but the representation of an object cursor in this form is
very large, because it converts the hex-encoded string into a JSON array
of numbers (bytes), stringifies that and Base64 encodes the result.

By decoupling `Page` and `Cursor`, we can define a new Cursor type for
Object cursors (which will BCS-encode the underlying type before
Base64 encoding it) with a more compact representation, and still take
advantage of the pagination framework.

Test Plan:

Behaviour preserving refactor:

```
sui-graphql-rpc$ cargo nextest run
sui-graphql-e2e-tests$ cargo nextest run -j 1 --features pg_integration
```