[GraphQL/Cursor] Object pagination
Ports `Object` to use the new cursor and pagination framework.  This
comes with a number of other supporting changes (which can be split
into their own PRs if helpful).

- Introducing a new `BcsCursor` that obfuscates cursor values using
  BCS instead of JSON (for a more compact representation).
- Binding extra values representing ObjectIDs as cursors for the test
  adapter, so we can implement object pagination tests.
- Renaming the existing `Cursor` to `JsonCursor` to differentiate it.
  We don't want to get rid of this entirely because it's much more
  convenient to work with (e.g. during testing, to create cursors from
  arbitrary structs).
- Factored out the intersection helpers used by
  `TransactionBlockFilter` into their own module.
- Implemented filter intersection for `ObjectFilter` in a similar style.

`Object::paginate`:

This is mostly implemented as-is in `db_data_provider`.  The main
difference is in how  the `owner_type` field is handled.  Namely, we
separately apply a bound on `owner_type` if one is supplied to the
function, and apply a bound to state that the type must be `Address`
or `Object` if an `owner` address is supplied.

Empirically this produces query plans with as good or better
complexity than the approach that used a single query on `owner_type`.

`DeprecatedObjectFilter`:

To split up the replacement of object loading APIs, the old
`ObjectFilter` has been retained as `DeprecatedObjectFilter` (the new
`ObjectFilter` changes the type for the `type_` filter).

`objectKeys`:

This PR also puts in an initial implementation of `objectKey`
filtering.  Note that rather than being an `AND` query, object key
queries are `OR`-ed with `objectId` queries and the union of both of
them is applied as a single filter in conjunction with other filters.
This is so that a single query could look for objects with known and
unknown versions. (The straightforward implementation would create a
query that performs the intersection of the `objectIds` and
`objectKeys` filters, which would make it useless to try and use both
at the same time).

This is carried through to the implementation of `ObjectFilter`
intersection, which correctly handles the case where one filter
specifies an ID, and another specifies an ID and version.

Test Plan:

New unit tests for `ObjectFilter` intersection:

```
sui-graphql-rpc$ cargo nextest run -- types::object
```

The following E2E tests highlight the differences in behaviour before
and after this change:

```
sui-graphql-e2e-tests$ cargo nextest run \
  -j 1 --features pg_integration --      \
  call/owned_objects.move                \
  object/pagination.move
```

The first one shows the "intersection" logic in action, and the second
shows cursor obfuscation.