fix: docs (#15802)
## Description 

This PR fixes the documentation error in Custom Transfer Rules. Which
resolves the issue where docs says `or it's locked, and only 0xCAFE can
transfer it.` and changes it to `or it's locked, and it can only be
transferred to 0xCAFE.`

## Test Plan 

This doesn't changes anything in the working code and doesn't needs any
testing.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

---------

Co-authored-by: ronny-mysten <118224482+ronny-mysten@users.noreply.github.com>