[GraphQL/Upcast] Remove asObject and asMoveObject upcasts
Extends `IOwner` interface for:

```
- `MovePackage`
- `MoveObject`
- `Coin`
- `CoinMetadata`
- `StakedSui`
- `SuinsRegistration`
```

Introduces `IObject` interface and implements it for:

- `Object`
- `MovePackage`
- `MoveObject`
- `Coin`
- `CoinMetadata`
- `StakedSui`
- `SuinsRegistration`

Introduces `IMoveObject` interface and implements it for:

- `Object`
- `MovePackage`
- `MoveObject`
- `Coin`
- `CoinMetadata`
- `StakedSui`
- `SuinsRegistration`

This allows, e.g. a `Coin` to be treated like a `MoveObject`, an
`Object` or an `Owner` without potentially multiple explicit upcasts.

This also requires changing the names of some existing fields to not
conflict with fields on their super class:

- `Coin.coinBalance` to deconflict with `Owner.balance`.
- `MovePackage.packageBcs` to deconflict with `Object.bcs`.
- `StakedSui.stakeStatus` to deconflict with `Object.status`.

Originally, the intention was to use `#[graphql(flatten)]` to
implement this, but this doesn't seem to work consistently in the
presence of recursive type definitions, so the approach is more
manual:

- Each interface is implemented as a combination of a GraphQL interface,
  and a Rust `*Impl` type that holds the re-usable implementations.
- Each GraphQL type that implements an interface has to contain copies
  of all the function signatures it implements in the common
  interfaces, that make a delegated call to the Rust `*Impl` type.
- `async-graphql`'s implementation of interfaces ensures that we don't
  miss an implementation.

This PR also cleans up some docs for consistency across all the shared
fields.

Test Plan:

Existing unit and E2E tests:

```
sui-graphql-rpc$ cargo nextest run
sui-graphql-e2e-tests$ cargo nextest run -j 1 --features pg_integration
```