[graphql] Fix health check error (#15863)
## Description 

The logger extension needs the `uuid` for the query as well as the `ip
address`, otherwise it will crash. This happens for the `health`
endpoint that does not call the `graphql_handler`.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Fixed health check error due to the health endpoint not setting the uuid
and socket addr.