[GraphQL/SuiNS] Introduce NameService (#15930)
Based on feedback from Manos that it is confusing to have the resolution
functions hanging off `SuinsRegistration`.

This slightly goes against an earlier recommendation I made to not
introduce functions on types that don't need `self` unless they are
constructors of some kind, but it is done to match the pattern we have
for queries.

Existing tests:

```
sui-graphql-rpc$ cargo nextest run
sui-graphql-e2e-tests$ cargo nextest run -j 1 --features pg_integration
```