[gql][consistent-reads][3/n] Implement consistent reads for Object and MoveObject (#15915)
1. Implement `RawQuery` version of the existing TypeFilter and
ObjectFilter logic
2. Note that when object_keys - object_id and object_version - are
provided, the WHERE condition gets dropped. This is so that we can
actually return data like the gas payment of a transaction. In other words, return the historical
3. test_adapter - interpolate cursors
4. extend paginate_raw_query again to accept a cursor_fn that transforms
each element's cursor. This is predominantly for Object, as the
checkpoint_sequence_number to be set on an object as context should come
from the method parameters, not the object itself.
5. Change Gas.gasPayment to a Connection of MoveObject, not sure if its
intended to be a MoveObject or we missed it

tx_address_objects.move
objects_pagination.move

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration