[gql][consistent-reads][x/n] Make cursor's checkpoint non-null, add context from the input and transaction connections down to ObjectRead. (#15996)
## Description 

Adds a `ConsistentIndexCursor` to the `consistency` module - we have a
good number of cursors that are just indices into a type's `Vec` field,
so this is a quick way to handle consistency for all of them in one go.

Sets `checkpoint_viewed_at` field on cursors to non-null. `DryRun` and
`Mutation` results now fetch the latest checkpoint sequence number so
that we can set `TransactionBlock` and `TransactionBlockEffects` to a
checkpoint value. This is so that pagination fields on types that do not
make a db roundtrip will have the correct checkpoint context on hand
already - thus we can set `checkpoint_viewed_at: u64` on fields like the
transaction calls of a `PTB`. Actually, it's likely the case that all
graphql types can have a non-null `checkpoint_viewed_at`, but this can
be revisited later.

Removes `fn cursor` from `Target` trait. Adds a
`paginate_consistent_indices` function - does exactly what
`paginate_indices` does, with the one difference that this function also
validates cursor consistency. The caller just needs to use `cursor.ix`
to index into the data's `Vec` field, and use `cursor.c` which has been
set to the correct `checkpoint_viewed_at`.

Shorten cursor fields to one letter to reduce size.

This leaves MoveModule and related types and Validator and related
types.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes