move-package: make `nested_deps_git_local` use sui repo for dep (#16023)
## Description 

Forgot this yesterday with the nits. It's related to this
https://github.com/MystenLabs/sui/pull/15867/files/3ee2e7f1df0cfea267a0b25896b46a5d6c61629b#r1466324788

## Test Plan 

How did you test the new or updated feature?

N/A

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes