[gql][consistent-reads][n/n] Implement consistent reads (#15913)
The top-level commit of a stack of PRs for ensuring that a `GraphQL`
query is consistent from its root to leaf. To do so, we thread a
`checkpoint_viewed_at: Option<u64>` field through each `GraphQL` type.
This field is populated when a type is resolved and when a cursor is
built. This field needs to be on cursors as well, so that when
paginating through a connection, we can lock the pagination queries to a
specific checkpoint in time.

A few queries to demonstrate:
```
{
  transactionBlock(digest:"digest"){
    sender {
      objects {
        nodes {
          address, version
        }
      }
    }
  }
}
```
The query is made at checkpoint 10, thus `checkpoint_viewed_at=10`. The
nested `objects` field inherits `checkpoint_viewed_at=10`.

```
{
  object(address: "address", version: 10) {
    owner {
      ... on AddressOwner {
        owner {
          objects {
            nodes {
              digest
            }
          }
        }
      }
    }
  }
}
```
Similarly, the nested owner and objects all inherit
`checkpoint_viewed_at=10`.

```
query {
  address(address: "...") {
    coins(type: "0x2::sui::SUI") {
      nodes {
        owner {
          ... on AddressOwner {
            owner { asAddress { balance(type: "0x2::sui::SUI") { totalBalance } } }
          }
        }
      }
    }
  }
}
```
If `checkpoint_viewed_at=10`, then each coin result will yield an owner
that inherits `checkpoint_viewed_at=10`. Subsequently, each balance
result reflects the owner's balance at checkpoint 10.

```
query {
  address(address: "...") {
    coins(type: "0x2::sui::SUI", cursor:"checkpoint=10") {
      nodes {
        owner {
          ... on AddressOwner {
            owner { asAddress { balance(type: "0x2::sui::SUI") { totalBalance } } }
          }
        }
      }
    }
  }
}
```
Some time later, the user makes another query using a cursor from the
first query. Now, `checkpoint_viewed_at=20`. When we hit the coins
connection field, `checkpoint_viewed_at` is set to `10`, and this value
is inherited by the rest of the children.

This stack adds consistent context to the remaining types, and completes
the consistency framework.
https://github.com/MystenLabs/sui/pull/16006
https://github.com/MystenLabs/sui/pull/15996
https://github.com/MystenLabs/sui/pull/15962
https://github.com/MystenLabs/sui/pull/15961
https://github.com/MystenLabs/sui/pull/15960
https://github.com/MystenLabs/sui/pull/15959

This stack implements historical queries for `Object` and its subtypes.
It also sets up a nascent framework for supporting consistent reads.
https://github.com/MystenLabs/sui/pull/15967
https://github.com/MystenLabs/sui/pull/15958
https://github.com/MystenLabs/sui/pull/15963
https://github.com/MystenLabs/sui/pull/15932
https://github.com/MystenLabs/sui/pull/15931
https://github.com/MystenLabs/sui/pull/15915
https://github.com/MystenLabs/sui/pull/15914

How did you test the new or updated feature?

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration