[move-ide] Added support for dot-call syntax (#16054)
## Description 

This PR adds support for dot-call syntax. In particular, `use fun`
declarations are supported:


![image](https://github.com/MystenLabs/sui/assets/1724397/f5f4ecce-3085-4e89-803e-f7d0d7408c83)

Also, method calls are supported:


![image](https://github.com/MystenLabs/sui/assets/1724397/44d860d4-6726-46de-9963-9fa3dbcdba4d)


 It required a few additions to the compiler, in particular:
- added `UseFuns` to `Sequence` and `ModuleDefinition` in the typed AST
so that it's available to the IDE visitor
- added optional method name to `ModuleCall` in the typed AST so that we
have it available when we are computing symbols over typed AST
- added `TypeName` to `UseFun` in the naming AST so that we can keep
track of both `SomeStruct` locations in the following code snippet:
```
    public use fun Self::foo as SomeStruct.f1;
    public use fun Move2024::M1::foo as SomeStruct.f2;
```
-

## Test Plan 

New tests have been added.