[move-2024] Correctly recompute breaks under named blocks (#16052)
## Description 

Due to a wildcard match, recompuing `loop` breaks did not proceed under
them in HLIR lowering. This fixes that and revises HLIR lowering to
always exhaustively match the forms when looking for breaks.

Also, the testing found cases where we had `annot((x: _|_): T)`, causing
freeze operations to fail. This modified freeze checking to pass when we
find `x: _|_` because that expression is unreachable and will be
removed, plus typing was fine with it.

## Test Plan 

A very long test for all these forms.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes