Authority overload monitor (#16050)
Roll forward #15981 

Previously, the PR was causing CI failure because `max_txn_age_in_queue`
was used in the config. Therefore, although `overload_threshold_config`
has a default, since `max_txn_age_in_queue` is specified, serde expects
the rest of config to show up in the config file. Because we don't have
a serde default set for the rest of the field, it crashes.

This PR added serde default for all the field in OverloadThresholdConfig
in addition to the original #15981 .

I manually verified that it can parse the config correctly when only
partial config is set.