[CLI] Add a number of improvements (#16068)
## Description 

This adds a number of improvements/features, I've organized them by
commit:
1. Named address resolution: This adds support for things like:
```
--assign my_addr @0xabc
--move-call "my_addr::my_module::my_func" ...
```
and this will resolve the `my_addr` occurence to `0xabc`.

2. Adds support for parsing Move calls without space between the type
arguments and the function name. This adds support for things like:

```
--move-call "std::option::is_none<u64>" "none"
```

Previously this would be a parse error.

3. This adds support for pointing at gas-budget/picker errors.
4. Adds key-store resolution for mnemonic addresses. So if I have a
`sad-tree` mnemonic for an address I can then do something like:
```
--assign s sad-tree
--move-call "s::..."
```
You can also use the mnemonic directly as an address:
```
--move-call "sad-tree::..." ...
```
Note that this commit also broadens the scope of what is an acceptable
identifier (to include `-`) so that we can capture default mnemonic
identifiers.

## Test Plan 

Tested out locally

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes