[NameService] Subdomain support (#15576)
IndexerV2 & GraphQL logic will be carried over on the next PRs.

- Adds checks for expiration on SLD names
- Adds subdomain lookup & expiration checks (for both node/leaf
subdomains)
- Changes error structure for easier support between the 3 crates that
use it

Added some unit tests and did some localnet testing for the resolution
aspect.
Not sure if there's a better way to test this (need advise!)
Will add some extra unit tests once we're good with the logic.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration