[move] Make most ICEs diagnostics (#16124)
## Description 

The compiler has a number of panics when things have gone wrong. This
tries to remediate a large number of them to return some sort of error
to allow compilation to continue, plus report it as a bug.

It's unclear if this will make development easier, but I think it is
overall a situation we should move toward to helps when users stumble
into these. If there are any specific panics we'd like to reinstate,
please call them out.

## Test Plan 

The test suite behavior should not change. 

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes