for convenience
ok i think this approach actually works? it ended up being quite simple. if empty response, just default to scan_prev and scan_next page. In fact, always default the bool prev_page and next_page to the scan. The actual cursor value is determined by whether we had an empty response or not. when pruning, it's not enough to just naively prune if cursor were specified, we need to check that the tips actually match - so if after, first element matches, if before, last element matches

update the corresponding, starting from last