[cherry-pick] Add version metric into new release (#16419)
## Description 

This PR adds a version to the prometheus metrics to allow us to
understand version and sha that the service is running.

## Test Plan 

Locally.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
This PR adds a version to the prometheus metrics to allow us to
understand version and sha that the service is running.