TxManager::enqueue does not need to return Result (#16450)
## Description 

To be consistent with all the other calls inside the function, where we
always use expect.
It's also a bad practice to ignore the return value of some calls.
Changing all of them to expect.
The core of this PR is in TransactionManager. All changes in other
functions are due to the function signature change.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes