[bridge 47/n] updates to limiter & get_token_transfer_action_status  (#16373)
## Description

This PR contains two major things:
1. updates to limiter - previously limiter's limiting direction is
incorrect. The goal is to limit on the receiving end (where money flows
out of the bridge) but we did in the opposite way. This PR fixes it
2. add a public function `get_token_transfer_action_status` so we can
use devInspect to get a transfer's current on chain status.

## Test Plan

* unit tests in move.
* off chain integration of `get_token_transfer_action_status` is in the
next PR for readability.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes