Deterministic Build using StageX (#16522)
## Description 

This uses the stagex toolchain (see: https://codeberg.org/stagex/stagex)
to build the`sui-node` container.

### What is stagex?

[stagex](https://codeberg.org/stagex/stagex) starts with 256 bytes of
x86 assembly and deterministically reproduces every dependency necessary
to build modern software. Every release of stagex is coordinated such
that only if two maintainers with a separate hardware stack are able to
reproduce the _entire_ toolchain, will the containers be published. As
such, stagex can be built locally to get the _exact_ same images that
exist in Docker Hub (when using Docker 25.x with containerd
snapshotter).

This pull request uses the stagex toolchain and the stagex Containerfile
workflow to build a deterministic OCI image. The image may be loaded to
Docker using:

```sh
# Import into Docker
env -C build/oci/sui-node tar c . | docker load

# Verify Sui loads by printing version
docker run sui-node
```

The choice of the OCI output format is to ensure reproducibility, as
Docker's image format has occasional quirks depending on the machine it
runs on, but we've yet to experience a hardware configuration that has
yet to reproduce using Docker 25.x and the containerd snapshotter.

Please note that the resulting image hash is dependent on the content
`ADD`ed in the Containerfile, so any modifications to any files that
aren't exempted in `.dockerignore` will result in a new image hash.

## Test Plan 

Run `sh docker/sui-node/build.sh`, and compare the built hash:

```
 => exporting to oci image format
 => => exporting layers
 => => exporting manifest sha256:d82d09740b6e1e49742d2190766e285868f5eb42177af95423a352107bd8e43f
 => => exporting config sha256:40c360768857f5e23f2361872a04d6a7d99aad17d488703e9a368b90f4dda00f
```

Run `docker system prune` to clear the build cache, and run `sh
docker/sui-node/build.sh` again.