[cli] Remove OnceCell as it seems to cause issues with test coverage (#16554)
## Description 

This PR changes the way TestCluster is created by removing the
`OnceCell` as it seems to be creating some issues with files not being
found when test coverage is ran. It is a temporary fix until we
understand better why we get SIGSEGV and SIGABR signals when tests run.

## Test Plan 

Existing tests.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes