fixup: Limit upgrade compatibility to known framework versions
Don't test compatibility up to the max protocol version, only test it
up to the max protocol version at the time the branch was cut.

Otherwise we need to also cherry pick new framework snapshots, and
changes to the framework, and therefore potentially also changes to
the compiler that was used to compile the framework.