Delete async when new wallet context in sui sdk (#16459)
## Description 

Fix #16458 

## Test Plan 



---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [x] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
The `WalletContext::new` method does not have any asynchronous
dependencies, and thus it does not need to be defined as an asynchronous
function. This PR is a breaking change that modifies the signature of
`WalletContext::new` into a synchronous function. Users that depend on
`WalletContext::new` will need to update callsites accordingly by
dropping the `.await`.

---------

Co-authored-by: Hgamiui9 <jinlong@cplus.zone>