[pick #16650][GraphQL/DataLoader] Epochs
## Description

Initial plumbing to use `Db` as the source for a `DataLoader` and an
implementation of `DataLoader` for `Epoch` to exercise it.

## Test Plan

This is a behaviour preserving refactor, so we use the existing tests:

```
sui$ cargo nextest run -p sui-graphql-rpc
sui$ cargo nextest run -p sui-graphql-e2e-tests --features pg_integration
```

@wlmyng also manually tested a query that would previously have done
multiple sequential Epoch reads, and found a 10x speed-up (3s to 300ms)
in local testing.