[gql] clean up config.rs (#16662)
Today we have two entrypoints, `start-server` and `from-config`. The
former is the standard way we launch the graphql service, by providing a
`--config` path to a `.toml` file. The latter is a yaml-based
entrypoint. Because we do not use this yaml config, this PR removes
yaml-related code to simplify. Also refactors `config.rs` to this order:
1. consts and struct defs
2. impl async-graphql's Object macro
3. impl
4. impl Trait

Note that `NameServiceConfig` has a `domain_type_tag` field that is
computed, so it probably won't play nicely with reading from a `.toml`
file. I'll address this in a follow-up PR.

Existing

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration