Add Bridge EVM Contract Project (#16449)
This adds the foundry project to the workspace.

The foundry project currently includes unit tests / fuzz tests with
plans to include invariance tests soon.

It is also intended that E2E bridge tests will deploy the bridge
contracts to an anvil node.

Note: Typically foundry leverages git modules to manage dependencies.
This is replaced with a build script found in the `sui-bridge` crate.