[pick #16755][GraphQL/E2E] Fix protocol version to 39
## Description
Make sure all existing E2E tests used a fixed protocol version (I
picked 39), so that when there are changes to a future protocol
version, they won't cause all the digests/addresses etc in these tests
to change.

Also updated the `update_all_snapshots` script to not update these
snapshots (should no longer be required) and therefore removed the
dependency on `psql`.

Added a section to the E2E test README to remind people to set a
protocol version. I opted for an approach where we set a protocol
version for each test rather than setting it once for all tests
because:

- The logic that controls which protocol version to use is embedded
  within the sui-specific init logic, but the entrypoint we use in the
  test runner is from the Move transactional test codebase (which is not
  aware of protocol versions, etc).
- We may want to set different protocol versions for different tests,
  in general, and so setting a blanket override or a hidden different
  default (other than the latest protocol version) seems liable to
  confuse people.

## Test Plan
Created a change to the framework (and a protocol version bump) on top
of this change and ran the tests -- they still passed, with no need to
change snapshots.