[Verifier][Chore] Extract Meter trait from Verifier
## Description

Pulling the `Meter` trait definition out of each version of the
verifier into a common place. This is so that this trait can be used
as part of the execution layer interface.

The overall motivation is so that we can supply a `Meter`
implementation that captures more data when run from the CLI, to
improve the quality of output from `sui client verify-bytecode-meter`.

## Test Plan

```
sui$ cargo build --bin sui
sui$ cargo simtest
sui$ env SUI_SKIP_SIMTESTS=1 cargo nextest run
```