Fix audit feedback on committee creation (#16909)
## Description

As reported by audit if a validator registers with an already used
pubkey for the bridge (whether maliciously or not) the creation of the
committee will fail and as such an end of epoch would fail attempting to
create the committee.
Moreover there is no way at the moment to flush the proposed committee
(in that respect, should we have an API for it?) and it would require an
emergency release to take care of the problem.
This should prevent the problem from happeing.

## Test Plan

Added unit test

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes