Check bridge routes on send token (#16960)
## Description

Audit found that `send_token` was not checking for valid routes.
Added the missing check

## Test Plan

Added unit tests.
Positive tests (no abort) for `send_token` are a bit trickier and will
be added when unit test review is done. They were not there so this is
not a regression in tests.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes