More audit, rename and invariants (#16924)
## Description 

More on audit feedback. Renaming as in the TODOs and invariants pulled
at the entry of the function.
`approve_bridge_message` has been renamed to `approve_token_transfer`
and invariants have been pulled all the way up to the entry. You'll
notice that because of that, the logic of the function changed a bit,
both in terms of errors and in terms of what gets executed
unconditionally.
I think it's fine, but please review 

## Test Plan 

Unit tests

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes