[CI] use seed specific to commit for Mysticeti simtest (#16985)
## Description 

When a simtest is flaky, unrelated PRs can encounter failures which
cannot be easily reproduced locally or in simtest nightly. It doesn't
seem productive to block the PR submission in this case.

## Test Plan 

CI

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes