[Types] Abstraction for defining built-in IDs and Packages
Main motivation for this change is to have a single source of truth to
drive the `is_system_package` predicate and a list of system packages.
This will be used by the GraphQL service to enumerate packages to
evict from its package resolution cache on epoch boundaries.

In the process, I ended up creating a macro to avoid some of the
duplication in this module (abstracts away defining all built-in IDs
as `AccountAddress` and `ObjectID`, and provides a unified way to
create the associated vanity IDs from a single `u16`).

## Test Plan

Existing simtests